+++ x86/src/main/scala/com/scalaAsm/x86/Instructions/INT.scala
@@ -18,7 +18,7 @@ object INT {
   implicit object int1 extends INT_1[imm8] {
     def get(x: imm8) = new Instruction {
       val operands = I[imm8](x)
-      val opcode = OneOpcode(0xCD.toByte)
+      val opcode = OneOpcode(0xCD)
       val modRM: Option[AddressingFormSpecifier] = Some(getAddressingForm1(operands))
       val opcodeExtension = None
     }
